requirements.txt
langchain
langchain-community
langchain-openai
langchain-experimental
streamlit
pandas
seaborn
matplotlib
pygwalker
jupyterlab
tiktoken
tabulate
python-dotenv
langchain-google-genai